pyproject.toml
[project]
name = "klangbild"
version = "0.1.0"
description = "4K audio visualizer for YouTube — waveform + cover image generator"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.10",
    "numpy>=1.26",
    "Pillow>=10.0",
    "tqdm>=4.66",
    "mutagen>=1.47",
]

[project.scripts]
klangbild = "klangbild.__main__:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["klangbild"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "mypy>=1.11",
    "ruff>=0.5",
    "types-tqdm>=4.66",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
